Update models.py
* Change expires time to a float to be correct data type

+++ models.py
@@ -32,7 +32,7 @@ class EpicTokenResponse(BaseModel):
     access_token: str
     token_type: str
     expires_in: int
-    expires: int
+    expires: float
     scope: str
